Refactor: Remove unnecessary calls to EGraph.getClasses (#4)
This PR is a refactor PR. It simplifies the logic, and also makes it
much faster, by removing unnecessary calls to the very expensive method
`EGraph.getClasses`.

This is accomplished by introducing a new "tag" which tracks the uses of
functions in each EGraph component. This allows only relevant objects to
be checked, sometimes giving significant savings. This PR also
eliminates the redundant `EMatcher` class.