Refactor EGraph and UF internals for clarity and performance (#11)
This PR updates the implementation of EGraph. It keeps a smaller set of
applications indexed by improving the way function applications are
canonicalized for implementing `updateCongruence()`. It skips congruence
maintenance for `not: true` and `eq: true` functions, because
applications of these functions don't generally benefit from congruence.

In addition, the responsibility of tracking object definitions is
removed from `UFSolver` and into `EGraph`.

These changes bring a small performance improvement:

* Current main abd0fa69a9ca8741e4174bc9619dbe9f0c793d1e
    * > Slowest: verify_tests.duplicated-assert-is-fast took 1248 ms
      > Done in 3.06s.
* 0cc3c54404b03b71f4f11898220bbfd9c4daa6af
    * > Slowest: verify_tests.duplicated-assert-is-fast took 1053 ms
      > Done in 2.84s.